Factors DerivedColumn out into its own module.
